+++ vscode/src/api/index/dependi-index-server/reports.ts
@@ -35,8 +35,8 @@ export enum Language {
   Rust,
   Golang,
   JS,
-  PHP,
   Python,
+  PHP,
 }
 
 const LanguageArray = [