+++ vscode/src/config.ts
@@ -5,18 +5,23 @@ export const DEPENDI = "dependi.";
 const WORKBENCH_ACTIONS = "workbench.action.";
 
 export enum Configs {
+  RUST_ENABLED = `rust.enabled`,
   RUST_INDEX_SERVER_URL = `rust.indexServerURL`,
   RUST_IGNORE_UNSTABLES = `rust.excludeUnstableVersions`,
 
+  NPM_ENABLED = `npm.enabled`,
   NPM_INDEX_SERVER_URL = `npm.indexServerURL`,
   NPM_IGNORE_UNSTABLES = `npm.excludeUnstableVersions`,
 
+  PHP_ENABLED = `php.enabled`,
   PHP_INDEX_SERVER_URL = `php.indexServerURL`,
   PHP_IGNORE_UNSTABLES = `php.excludeUnstableVersions`,
 
+  GO_ENABLED = `go.enabled`,
   GO_INDEX_SERVER_URL = `go.indexServerURL`,
   GO_IGNORE_UNSTABLES = `go.excludeUnstableVersions`,
 
+  PYTHON_ENABLED = `python.enabled`,
   PYTHON_INDEX_SERVER_URL = `python.indexServerURL`,
   PYTHON_IGNORE_UNSTABLES = `python.excludeUnstableVersions`,
 
@@ -55,22 +60,27 @@ export enum Configs {
 export const Settings = {
   version: "0.0.1",
   rust: {
+    enabled: true,
     index: "",
     ignoreUnstable: false
   },
   npm: {
+    enabled: true,
     index: "",
     ignoreUnstable: false
   },
   php: {
+    enabled: true,
     index: "",
     ignoreUnstable: false
   },
   go: {
+    enabled: true,
     index: "",
     ignoreUnstable: false
   },
   python: {
+    enabled: true,
     index: "",
     ignoreUnstable: false
   },
@@ -110,22 +120,27 @@ export const Settings = {
     const config = workspace.getConfiguration("dependi");
 
     // fill in the settings
+    this.rust.enabled = config.get<boolean>(Configs.RUST_ENABLED) ?? true;
     this.rust.index = config.get<string>(Configs.RUST_INDEX_SERVER_URL) || "https://index.crates.io";
-    this.rust.ignoreUnstable = config.get<boolean>(Configs.RUST_IGNORE_UNSTABLES) || true;
+    this.rust.ignoreUnstable = config.get<boolean>(Configs.RUST_IGNORE_UNSTABLES) ?? true;
 
+    this.npm.enabled = config.get<boolean>(Configs.NPM_ENABLED) ?? true;
     this.npm.index = config.get<string>(Configs.NPM_INDEX_SERVER_URL) || "https://registry.npmjs.org";
-    this.npm.ignoreUnstable = config.get<boolean>(Configs.NPM_IGNORE_UNSTABLES) || true;
+    this.npm.ignoreUnstable = config.get<boolean>(Configs.NPM_IGNORE_UNSTABLES) ?? true;
 
+    this.php.enabled = config.get<boolean>(Configs.PHP_ENABLED) ?? true;
     this.php.index = config.get<string>(Configs.PHP_INDEX_SERVER_URL) || "https://repo.packagist.org";
-    this.php.ignoreUnstable = config.get<boolean>(Configs.PHP_IGNORE_UNSTABLES) || true;
+    this.php.ignoreUnstable = config.get<boolean>(Configs.PHP_IGNORE_UNSTABLES) ?? true;
 
+    this.go.enabled = config.get<boolean>(Configs.GO_ENABLED) ?? true;
     this.go.index = config.get<string>(Configs.GO_INDEX_SERVER_URL) || "https://proxy.golang.org";
-    this.go.ignoreUnstable = config.get<boolean>(Configs.GO_IGNORE_UNSTABLES) || true;
+    this.go.ignoreUnstable = config.get<boolean>(Configs.GO_IGNORE_UNSTABLES) ?? true;
 
+    this.python.enabled = config.get<boolean>(Configs.PYTHON_ENABLED) ?? true;
     this.python.index = config.get<string>(Configs.PYTHON_INDEX_SERVER_URL) || "https://pypi.org/pypi";
-    this.python.ignoreUnstable = config.get<boolean>(Configs.PYTHON_IGNORE_UNSTABLES) || true;
+    this.python.ignoreUnstable = config.get<boolean>(Configs.PYTHON_IGNORE_UNSTABLES) ?? true;
 
-    this.vulnerability.enabled = config.get<boolean>(Configs.VULS_ENABLED) || false;
+    this.vulnerability.enabled = config.get<boolean>(Configs.VULS_ENABLED) || true;
     this.vulnerability.ghsa = config.get<boolean>(Configs.VULS_GHSA_ENABLED) || false;
     this.vulnerability.osvBatch = config.get<string>(Configs.VULS_OSV_BATCH_URL) || "https://api.osv.dev/v1/querybatch";
     this.vulnerability.osvSingle = config.get<string>(Configs.VULS_OSV_URL) || "https://api.osv.dev/v1/query";