+++ vscode/CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 
 All notable changes to the "dependi" extension will be documented in this file.
+
+## 0.7.4
+
+### New Features
+
+- PHP support added [PR #21](https://github.com/filllabs/dependi/pull/21)
+
+- Language specific enable/disable settings added [Issue #11](https://github.com/filllabs/dependi/issues/11)
+
+- Fetching dependencies notification silenced to status bar spinner [Issue #34](https://github.com/filllabs/dependi/issues/34)
+
 ## 0.7.3
 
 ### Bug Fixes
@@ -40,9 +51,7 @@ All notable changes to the "dependi" extension will be documented in this file.
 ### Bug Fixes
 
 - General bug fixes.
-
 - Fixed go.mod decorator bug ([#2](https://github.com/filllabs/dependi/pull/2))
-
 - Fixed retry command bug ([9f105dc](https://github.com/filllabs/dependi/commit/9f105dc347d66d33acf9e63c4c09d96206ae6e4c))
 
 ### Other Changes